requirements.txt
beautifulsoup4
flask
matplotlib
numpy
openai
pandas
plotly
prophet
python-dotenv
reportlab
scikit-learn
seaborn
selenium
sentence-transformers
shap
transformers
vaderSentiment
xgboost
